refactor: mv g2g type defs
