+++ packages/app/src/components/Admin/G2GDataTransferStatusIcon.tsx
@@ -1,36 +1,9 @@
 import React, { type ComponentPropsWithoutRef } from 'react';
 
-/**
- * Get the union type of all the values in an object, array or array-like type `T`
- * @see {@link https://github.com/piotrwitek/utility-types/blob/df2502ef504c4ba8bd9de81a45baef112b7921d0/src/mapped-types.ts#L589-L597}
- */
-type ValuesType<
- T extends ReadonlyArray<any> | ArrayLike<any> | Record<any, any>
-> = T extends ReadonlyArray<any>
- ? T[number]
- : T extends ArrayLike<any>
- ? T[number]
- : T extends object
- ? T[keyof T]
- : never;
+import { G2G_PROGRESS_STATUS, type G2GProgressStatus } from '~/interfaces/g2g-transfer';
 
 /**
- * G2G transfer progress status master
- */
-const G2G_PROGRESS_STATUS = {
-  PENDING: 'PENDING',
-  IN_PROGRESS: 'IN_PROGRESS',
-  COMPLETED: 'COMPLETED',
-  ERROR: 'ERROR',
-} as const;
-
-/**
- * G2G transfer progress status
- */
-type G2GProgressStatus = ValuesType<typeof G2G_PROGRESS_STATUS>
-
-/**
- * Props for {@link G2GDataTransfer}
+ * Props for {@link G2GDataTransferStatusIcon}
  */
 interface Props extends ComponentPropsWithoutRef<'i'>{
   status: G2GProgressStatus;
@@ -39,7 +12,7 @@ interface Props extends ComponentPropsWithoutRef<'i'>{
 /**
  * Icon for G2G transfer status
  */
-const G2GDataTransferStatusIcon = ({ status, className, ...props }: Props): JSX.Element | null => {
+const G2GDataTransferStatusIcon = ({ status, className, ...props }: Props): JSX.Element => {
   if (status === G2G_PROGRESS_STATUS.IN_PROGRESS) {
     return (
       <i className={`fa fa-spinner fa-pulse fa-fw ${className}`} aria-label="in progress" {...props} />
